fix bug

+++ app/Http/Controllers/CalendarController.php
@@ -558,7 +558,7 @@ class CalendarController extends Controller
 				$room_name = $room->name;
 				$order->room_id = $room_id;
 				if($person > 0){
-					// if(count($service_provider_list) - $no_specific_amount < $person) throw new Exception("該時段師傅數不足 請重新選擇", 1);
+					if(count($service_provider_list) - $no_specific_amount < $person) throw new Exception("該時段師傅數不足 請重新選擇 (error 1)", 1);
 					$order->serviceProviders()->detach();
 					$order->person = $person;
 					$service_provider_list = ServiceProvider::with(['leaves' => function ($query) use ($start_time, $end_time) {
@@ -612,7 +612,7 @@ class CalendarController extends Controller
 					$order->room_id = $room_id;
 				}
 				if($person > 0){
-					// if(count($service_provider_list) - $no_specific_amount < $person) throw new Exception("該時段師傅數不足 請重新選擇", 1);
+					if(count($service_provider_list) - $no_specific_amount < $person) throw new Exception("該時段師傅數不足 請重新選擇 (error 2)", 1);
 					$order->serviceProviders()->detach();
 					$order->person = $person;
 					$service_provider_list = ServiceProvider::with(['leaves' => function ($query) use ($start_time, $end_time) {
@@ -639,9 +639,6 @@ class CalendarController extends Controller
 						$service_provider_name = $service_provider_name.$service_provider->name." ";
 					}
 
-				}
-				else{
-					// if(count($service_provider_list) - $no_specific_amount < $order->serviceProviders()->count()) throw new Exception("該時段師傅數不足 請重新選擇", 1);
 				}
 			}
 			$order->start_time = $start_time;
